requirements.txt
amqp==2.5.0
appdirs==1.4.3
asgiref==3.2.3
asn1crypto==0.24.0
asyncio==3.4.3
attrs==19.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
cached-property==1.5.1
cachetools==3.1.1
celery==4.2.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
defusedxml==0.6.0
deprecation==2.0.6
diff-match-patch==20181111
Django==2.2.8
django-admin-list-filter-dropdown==1.0.1
django-admin-rangefilter==0.3.11
django-allauth==0.38.0
django-appconf==1.0.3
django-axes==4.4.0
django-bulk-update==2.2.0
django-compat==1.0.15
django-cors-headers==2.4.0
django-extensions==2.1.6
django-filter==2.2.0
django-guardian==1.5.0
django-hijack==2.1.10
django-hijack-admin==2.1.10
django-imagekit==4.0.2
django-import-export==1.1.0
django-ipware==2.1.0
django-jalali-date==0.2.9
django-js-reverse==0.8.2
django-model-utils==3.1.2
django-multiselectfield==0.1.8
django-redis==4.12.1
drf-yasg==1.17.1
django-simple-history==2.7.2
-e git+https://github.com/altostratous/django-simple-history-field-track.git@dacaaedd212e57f17c9152d7dcbcb2e68c676c59#egg=django_simple_history_field_track
django-telegrambot==1.0.1
django-timezone-field==3.0
django-tinymce==2.7.0
django-watson==1.5.3
django-widget-tweaks==1.4.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
idna==2.7
isodate==0.6.0
itypes==1.1.0
google-api-python-client==1.7.11
google-auth==1.8.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.7
isodate==0.6.0
itypes==1.1.0
jdatetime==3.0.2
jdcal==1.4.1
Jinja2==2.10.1
jsonfield==2.0.2
jsonschema==3.2.0
kombu==4.2.1
lml==0.0.9
lxml==4.2.5
MarkupSafe==1.1.1
numpy==1.15.2
oauthlib==3.0.1
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.14
packaging==19.0
pilkit==2.0
Pillow==5.3.0
psycopg2==2.7.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.7
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.0
python-crontab==2.3.5
python-dateutil==2.7.3
python-decouple==3.1
python-telegram-bot==11.1.0
python-zeep==3.1.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
raven==6.9.0
redis==3.5.3
requests==2.19.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.1.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
tablib==0.13.0
texttable==1.6.1
ua-parser==0.8.0
uritemplate==3.0.0
urllib3==1.23
user-agents==1.1.0
uWSGI==2.0.17.1
vine==1.3.0
git+https://git@github.com/BE360/django-plus.git
git+https://github.com/BE360/django-import-export.git
git+https://github.com/BE360/django-admin-persian-rangefilter.git
git+https://git@github.com/BE360/web-decouple.git
git+https://git@github.com/BE360/dictator.git
xlrd==1.2.0
xlwt==1.3.0
django-admin-multiple-choice-list-filter==0.1.1
freezegun==1.0.0
factory-boy==3.0.1
Faker==4.0.0
tblib==1.7.0
